Add MerkleTreeHandler
Originally this was intended to be an endpoint into a local JVM
process that builds Merkle trees of the local indexes but in the name
of prototyping evolved into an iterator over the vector clock data
stored in the index so that a Merkle Tree may be built on the Erlang
side.  In the future Joe's leveldb-based hash tree could be ported to
Java.