Build covering set of nodes for distributed searching
To perform a search yokozuna will need to generate a covering set of
nodes (shards) to pass to the Solr distributed querying API.  For now
the Riak Core coverage code will be used but it isn't ideal since it
does not try to minimize the set of nodes used.  In fact, I think it
will always return all nodes in the cluster which means Solr will
encounter duplicate doc ids.  According to the wiki duplicates will be
removed but how they are removed is non-deterministic.