Add partition and overlap filtering to search
Since Yokozuna replicates documents the shards will have overlapping
data.  This overlapping data must be filtered out for the search
results to be correct.  The first step is to filter the shards by the
list of partitions produced by the coverage plan.  In some cases this
will result in the unique set of documents but in most cases, and in
the default of 64 partitions and replication of 3, it will produce a
slight overlap.  To guarantee no overlap the second step is to
determine where the overlap occurs and then add an additional filter
on one of the overlapping partitions.  This filter is another list of
partitions.  For a doc to be included it's first owner must be in this
list of partitions.

These two filters are applied as a Solr _filter query_.  This method
was chosen because it doesn't affect the score and the results are
cached.  The cache overhead is a bit per document so 1 megabyte of
memory is good for over 8 million documents.

Two additional fields are added to the document.

1. `_pn` - _Partition Number_: The owning partition of this document.

2. `_fpn` - _First Partition Number_: The first owning partition in the
                preflist for this document.

The first field is used for the initial partition filter.  It allows
Yokozuna to filter "only docs owned by these partitions."  The second
field is used for the fine-grained filtering to allow Yokozuna to
filter "only docs owned by this partition and whose first owning
partition is in this list."

For example given the default ring size of 64 and replication of 3 the
following ASCII diagram shows a possible coverage scheme.  The `X`
marks a partition in the coverage plan and the hyphen trail leading up
to it is the range of partitions it stores replicas for.  This scheme
starts with partition 2 (ring starts at 1) and jumps by 3 for a
sequence of `[2, 5, 8, 11, ..., 53, 56, 59, 60, 63]`.  Notice that one
of the jumps, from 59 to 60 is not a distance of 3 but of 1.  This is
because the replication factor of 3 does not divide evenly into the
ring size of 64.  This means there is a remainder of 1 and thus
overlap.  In this case 60 overlaps with 59.  The partitions they have
in common are 59 and 58.  This means that for partition 60 only
documents with a first partition owner of 60 should be chosen.  This
is what the `_fpn` field is used for.

         LPN: 1 | 2 | ... | 56 | 57 | 58 | 59 | 60 | ...
    Coverage: ----X   ...   -X   -----------X   -X   ...

    LPN = Logical Partition Number

For the given example above Yokozuna would create the following filter
query.

    fq=_pn:P2 OR _pn:P5 ... OR (_pn:P60 AND (_fpn:60)) OR _pn:63