Store indexes in the ring and sync during ring events
An index is a cluster-wide resource.  If the create index API is
called on one node then it should be called on all nodes--even if they
aren't part of the cluster during the time of the call.  A metadata
entry in the ring is used to store the list of indexes.  During a ring
event the local list of lindexes is checked against that in the ring.
The delta is calculated and missing indexes are added.  Removal is not
yet implemented.

An internal index named `_yz` has been added.  Currently, it's only
use is to determine if Solr is up by pinging the `_yz` index.
However, it may be useful to use this index as a place to store data
internal to Yokozuna or Riak.

Finally, some changes were made to the process that starts Solr.  The
current working directory was changed to be the same as Riak's start
script, the lib dir is now passed at startup, and the solr log now
resides in same location as Riak logs, and make sure to wait until
Solr is fully started or things break when a ring event is handled.