Spitballing ideas for better indexing
There are three ways in which an object may be modified: 1)
write/update by client, 2) handoff, 3) read-repair.  A post-commit
hook only handles the first case.  By adding an "object modified" hook
to the vnode all 3 cases can be handled in the same manner.  After all
in every case the ulitmate goal is to update the object and when the
object is updated it's indexes should be updated as well.