Add new index hook to run during obj modification
A post-commit hook allows the index to be updated during an oject
write performed by a client.  But it will not handle the cases where
an object is modified by handoff or read-repair.  A solution is to
create a hook that is run in the vnode _any time_ an object is
modified.  This way an object's indexes are updated in all scenarios.