Periodically remove non-owned data
Over time partition data may move from node to node because of node
removal/addition or because of hinted handoff.  Add a periodic tick
event which searches out index data not owned by the local node and
removes it.

This will remove indexes for fallback data but that is okay because
fallback partitions are never queried.  In the future indexes
shouldn't be written for fallback data.