Wait for index creation on first handoff write
When a new node joins the cluster there is a race between handoff and
index creation.  The index creation is triggered by a ring event and
may take several seconds to complete.  Handoff is triggered by a ring
event and the vnode manager tick.  If using a forced join the handoff
can start quick enough that the new node hasn't finished creating all
the indexes.  If an index isn't ready then the hook will fail to index
the object.  Anti-entropy will converge the index with the object but
that may not happen immediately.

This solution waits for the index on the first handoff write for up to
`max_wait_for_index_seconds`.  This wait will only occur once per
vnode regardless if the index is ready or not after the max time has
expired.  In that case object indexing will fail until the index is
ready just as it did before this patch.  This simply gives the index a
chance to catch up.

In the best case the index will be ready before max wait and all
objects will be indexed correctly.  In the unfortunate case the index
will take a little longer and some objects will fail to index but
anti-entropy will fix that at some point.  In the worst case the index
will completely fail and all objects will fail to index.  This is
desired behavior because even though the indexes won't exist the raw
data is still making it to disk.  Once the problem with the index is
solved anti-entropy will create the indexes that were missed.