Make index API less confusing
Replace `add_to_ring` with `create` and previous create with
`local_create`.  The create API is used to enter the index name in the
ring, this is then gossiped causing ring events triggering a call to
local_create on each node in the cluster.  The local_create API
handles creation of the Solr Core which backs the index.