Delete doc on object reap
The are two stages to an object delete in Riak KV.

1. Write a tombstone value.

2. After a period of time, reap the object.  I.e. backend delete.

Step 1 will cause a new doc to be written with no text and thus it
won't show up in a query regarding the content of an object.  But the
document will still exist.

This commit hooks into step 2 so that the Solr docs are reaped along
with their object counterparts.  At this point the Solr doc should not
resolve for any query, even by id.