Add extractor support
The purpose of an extractor is to take the riak_object value and
convert it into a list of fields which can be used in a solr doc.  The
extractor used is based on the media type of the object.  I'm punting
on siblings for now, the head is taken from the list of values.