Remove riak_kv and ibrowse from deps
Technically, Yokozuna makes calls to functions in ibrowse, riak_kv,
and other application that come with riak_kv, such as riak_core.  But
none of this is needed at compile time, and Yokozuna releases are made
from the riak repository, not here.  So there is really no need to
have them as deps.  It only slows down compile.  It's also breaking
travis CI right now because of erlang_js.

The dialyzer targets will have to be adjusted but for now they will be
broken.