Add script to build custom Solr release
This script is used to build custom Solr releases for use with
Yokozuna.  Currently Yokozuna relies on a custom patch to distributed
search.  Rather than require people to build this release during the
Yokozuna build this script will be used to create a custom release
which will be uploaded to Basho S3 and downloaded by the Yokozuna
build script.  This will simplify and shoten the time it takes to
build Yokozuna.