Handle index content flag transitions
If the index content flag transitions from off->on and data has
already been written to the bucket then delete all data under the
default index for that bucket and clear the AAE trees.

If the transition is from on->off delete all data in the bucket's
index and clear the AAE trees.