Auto-add yz fields and uniqueKey
Automatically add the schema config needed for Yokozuna to function
correctly.  This will mutate the user's schema, stripping comments,
changing whitespace in some places, and redifining `uniqueKey`.  For
this reason I wonder if it would be better to simply reject
non-conforming schemas and let the user copy/paste the necessary
config.