Add bad schema test
Add a test that uploads a schema with a bad classname.  This will
cause the Solr Core CREATE call to fail which will cause an exception
to be thrown in yokozuna eventually leading to the death of the
cluster.

The next commit will protect Yokozuna from crashing in this case.