Don't crash on bad schema
It is fairly easy to create a schema that will cause a runtime error
in Solr when trying to create the core.  Currently Yokozuna causes the
entire cluster to shutdown in this scenario because yz_events reaches
max restart attempts.  This patch makes index creation a best effort.
It will periodically retry.  If the schema is bad the user can upload
a new one which will be used on the next retry.  The riak and solr
logs will be loud about failure and should have enough info to correct
the problem.