Remove clean_value from JSON extractor
The clean_value function is a relic from when the doc generated from
the extractor was being converted to XML to send to the Solr update
endpoint.  We have since switched to sending updates as JSON and thus
mochijson2 is used for encoding.  This handles Erlang terms as values
just fine and there is no need for special treatment of numbers.