Don't pass any JVM args as default
If a user doesn't specify solr jvm args in app.config then the default
should be to pass no args.  I think this is the least surprising
behavior to have.  We can still ship pre-canned defaults in the
app.config such as larger heap but the fallback defaults should be
none.