Add yz-fprof header for search resource
Allow user to pass `yz-fprof: <filename>` header to the search
end-point.  This will invoke fprof analysis and save it in the CWD of
the Riak process under the name `<filename>.analysis`.  This analysis
can be expensive so its primary use is development.  It may be called
in production but should be done sparingly and only when necessary.