Fix extractor resource
A regression was introduced in the last patch to the extractor
resource: it breaks if the `yz-extractor` header is not passed.  This
patch fixes that and adds a test.  The test was also changed to use
only 1 node since there is nothing about it that is distributed.  This
helps cut down on testing time.