Add M:1 support for Bucket:Index mapping
Allow many buckets to be indexed under one index.  This removes
the index flag and replaces it with a `yz_index` property which
is used to store the name of the index for which the bucket
should be indexed under.  If this property is a) non-existant or
b) equal to the default index then the bucket will be consisdered
a non-indexed bucket and the objects' values will not be indexed.

The HTTP API to create an index no longer automatically installs
a bucket property.  For now 2 calls will be required.

Still need to address index deletion.  Right now an index could
be deleted when buckets are still using it.

* Update riak tests to use new index API.  Added `wait_for_index`
  function that can be used instead of sleeps.

* `yz_flag_transitions`

  * Add 1s sleep after `load_data` to give soft-commit change to commit.

  * Assert the return value from `wait_for_aae`.

* `yz_index_amin`

  * Use smaller ring to make test faster.

  * Add 2s sleeps to prevent ring/metadata races.  Revisit this soon.

  * Put assert in `check_list` so on failure you can see actually values.

* `yz_languages`

  * Use smaller ring.

  * Use 1 node cluster since this test isn't really about clustering.

  * Use `wait_for_index` instead of sleeps.

* Ignore unused vars in `yz_pb`.

* Use `wait_for_index` instead of sleeps in `yz_siblings`.