Fix `sync_data`
I forgot to update the `sync_data` function for the new one-to-many
indexing feature.  When removing an index association from a bucket a
query delete must be sent to the index (not an index with the same
name as the bucket) to remove all documents with a matching bucket
(`_yz_rb`) field.

Send a commit after the query delete.  During testing I found that
some or all of the trees would get rebuilt before the query delete
took affect (the default 1s soft-commit).  Thus the test would fail
because it would have to wait for the trees to expire which defaults
to a week.  Instead, I thought it wouldn't hurt to run a commit after
the delete.

Added an additional check to the flag transition test to verify that
on adding a flag all entries are deleted from the default index.