Prevent Yokozuna and Riak Search conflicts
When Riak Search is enabled Yokozuna must take care not to take over
endpoints that it listens to.

* Don't register search PB services.

* Don't listen on the 'solr/<index>/select' endpoint.

Eventually a function will be added so that during migration a user
can tell Yokozuna to gracefully take over these interfaces from Riak
Search.  But the default should be to let Riak Search take preference
so rolling upgrades work.