Check for undefined solr port process
The main purpose of this commit was to avoid calling
`proplists:get_value` on an undefined port (Erlang port, not socket).
This would happen when the Solr process has a non-zero exit,
e.g. because the specific socket port number is already bound.

Some additional fixes I made along the way:

* Move `getpid` function in API section where it belongs.

* Fix the exit status match in `handle_info`.

* Check for a regular port exit msg in `handle_info`.

* Only close the port in `terminate` if the pid is not `undefined`.