Move Solr proc server under yz_sup
Move the solr proc server under the yz_sup so that max restart
frequency will actually be reached and shutdown the Riak node.  The
thinking is that if Solr cannot run then Riak should shutdown to avoid
missing index data and other badness.  That is, better to have a
downed node than one that is partially up.