Use rebar's git vsn for app version tagging
Replace the semver plugin with rebar's built-in `git` vsn support.
This will automatically use the tag value if building from the same
commit of a tag.  Otherwise it will use latest tag value + num commits
since + commit hash.  This is useful because it means every commit
generates a new application vsn and there is no chance of having the
same vsn number have different code associated with it.