First pass of Riak Search migration test
This test verifies that Riak Search data can be migrated simply by
letting AAE repair all the missing indexes.  There are many TODOs to
address to make this test even better.  But this is a good first
start.

The verify script now requires 2 arguments.  The dir of a previous
Riak build must be passed which the test will use as the version to
migrate from.