Fix repair count
Only count "full" repairs towards total count.  A `full_repair` is one
that actually repairs a missing or invalid Solr document.  A
`tree_repair` is one that only updates the Yokozuna hashtree because
the `BKey` is not stored in an indexed bucket.  In that case only the
tree should be updated to keep AAE happy.  A `failed_repair` is one
where the repair failed.  These can be ignored as the next exchange
will retry.