lookup index from bucket for mapred
Since yokozuna does not map buckets to indexes 1:1, it is necessary to
lookup the index to use when a search query on a bucket is specified.

This patch also supports both {index, Index} (as the search inputs will
use) and {struct, [{<<"index">>, Index}]} (as modfun inputs must use) to
allow the user to specify an index instead of a bucket.