git commit fix solr start timeout & verify to reset nodes correctly
If verify.sh builds stagedevrel, then the lib/yokozuna directories in the
dev nodes are symlinks to the yokozuna built in the work space. Since they
are symlinks, their contents are not put under git control. Thus, the git-reset
between tests does not replace the the solr.war deleted in yz_solr_start_timeout.

This patch uses devrel instead of stage devrel to make the git-reset work.