Use newer style stats
After discussing with Joseph Blomstedt I learned that the Riak KV stat
module has a lot of extra cruft for legacy reasons.  He pointed out to
me that the Riak Pipe stats module would serve as a better base for
Yokozuna stats.  It makes use of a newer style stats layout that
supposedly makes things easier.  I went ahead and mimicked it as best
as possible.  The only major exception is I have left in the sidejob
usage because some Yokozuna stats will fire on the KV VNode and we
want to protect them from stats work as much as possible.