a better way to make solr startup "fail" for testing
Being unable to use stagedevrel for testing was inconvenient, so this
patch brings it back.

Instead of deleting solr.war, the test configures the YZ data directory
to a well-known string that causes yz_solr_proc:build_cmd to return something
other than jetty/solr/java.

A catch was also added to the port_close command in terminate, because the
fake-solr (grep) dies much faster than the real solr, so that port_close call
will fail, obscuring the reason for the process exit.