Remove entropy data timestamp
The entropy data timestamp is a holdover left from the very early days
of Yokozuna.  The 3rd commit to be exact.  The idea was to allow only
building a tree from data up to a certain time.  This was to prevent
spurious repair in a non-quiescent cluster.

However, the AAE system in KV doesn't have this feature and thus it is
never used.  It is simply dead code taking up time and space.  Time to
say goodbye.