Don't store all _yz fields
The entropy data, node, partition and first partition fields do not
need to be stored.  They are only used for filtering by Yokozuna and
are of no immediate use to the user.  The `_yz_id` field is also of no
use but Solr's distributed search requires the id field to be stored.

By not storing these fields a lot of space is saved in the index.  In
one benchmark disk usage was reduced by 25%.