Add ability to reload indexes
Add the ability to reload indexes on a running cluster. This is useful
for loading schema changes on the fly. E.g. adding a field to the
schema.

The index reload call is a blocking call that broadcasts cluster-wide.
The caller should always check the return value. If any errors occurred
then the reload should be retried until all nodes report success.
Otherwise you could end up with different version of the schema across
the nodes.