Use more common index setup
The most common index setup is to have one bucket type, like `data`,
with each bucket underneath it pointing to its own index. Use this
style in the essential test to avoid oddities like using buckets named
`b1`.

Remove the siblings index creation as it is no longer used in this
test.