Remove compile/unit from verify script
This step verifies that Yokozuna compiles and passes unit test but
compiling top-level Riak already verifies that Yokozuna compiles. This
means a ton of extra network traffic and time is being spent just to
verify the unit tests. Remove this step and instead focus on running the
riak tests more quickly.