Change index exists check to use on-disk information
Add a function to get the lists of indexes created on disk. When
determining if an index exists use this information rather than a ping
request. This is more reliable because a ping request will fail if Solr
is down but the index actually exists. That causes unnecessary
error messages because Yokozuna will try to re-create indexes when
they already exist while Solr is down.