Change index exists to check 3 conditions for existence
It's not enough to just check that the index is in the official list and
exist on disk. The index might have a bad schema which keeps it from
starting. Without this patch the schema admin test will fail for just
that reason.

When Solr is up then the index should meet all three conditions below.
When Solr is down only the first two conditions must be met to be
considered as existing.

1. The index is in the "official list" (currently stored in ring).

2. The index has a dir under root data dir.

3. The index can be pinged via HTTP.