Perform handoff check only when Yokozuna is enabled
The handoff check only needs to be performed when Yokozuna is enabled.
Checking it in all cases doesn't break anything at the moment but if the
code was changed in the future it could cause KV handoff breakage when
Yokozuna is disabled. In that case, don't put KV in harms way. Don't
perform the check.