Don't make metadata call when service is down
Check for metadata convergence only when the Yokozuna service is running
on the other side.

If the service is not running on the other side then it indicates mixed
cluster or that Solr has not finished starting on the remote node. In
both cases there is no need to block handoff on metadata convergence
because there is no way to index the data anyways. That is, the metadata
check is currently only there to make sure that the `search_index`
properties have been propagated so that indexing events are not missed.
If Yokozuna doesn't exist or Solr is not up then no need to worry about
bucket properties since the indexing will fail anyways.