Add comment about AAE + ring resize
While going through an old KV commit I notice a try/catch was added
in the start_exchange method to avoid a crash of the entropy manage in
the case where exchange is attempted on an index that doesn't exist
anymore because of ring resizing. I disagree with the use of the
try/catch because it is too granular and it is sloppy. It assumes that
if an exception occurrs it is only for that reason. Interactions between
disparate systems should not be handled by try/catch + logging. We
should think carefully about how they interact and make sure to put in
the appropriate measures so they do not step on each others feet rather
than doing defensive programming. In this case the `maybe_exchange`
function could check to make sure that the index (partition) still
exists before performing the exchange.

basho/riak_kv@006f820665ecd779f7ef344454cfa91c8acc1494