Add verification of orphan postings cleanup by AAE
There can be scenarios where Solr contains postings for KV objects which
no longer exists. This test verifies that AAE cleans those up as it
should. The test bypasses the regular KV get and just creates postings
in Solr and then verifies that AAE calls the repair function the correct
number of times.