Add N value to index info
Add N value to index info. Up until this point Yokozuna contained a bug
in the coverage plan code where it pulled the N value from the bucket with
the same name as the index. This incorrectly assumed a) that the index
had a bucket associated with the same name and b) the index had any
buckets associated with it at all. After thinking on it for some time it
became clear to me that each Yokozuna index should have it's own N
value. This will be used to build the proper coverage plan. However,
this means that any buckets that write to this index should also have
the same N value as the index. If the bucket's N value is smaller than
the coverage plan will give nondeterministic results because it will
assume more replicas than actually exist. If the bucket's N value is
larger then the coverage plan will be less available than it could be.