Download custom jars from s3
Both Solr and Yokozuna's custom class files must be built with JDK 1.6
so that it will work for JRE 1.6 and greater (official Solr 4.x builds
against 1.6 so I'm following their lead). Instead of building these jars
fresh each time, leaving them at the mercy of whatever javac happens to
be on the official build machine, download them from s3. This way the
jars can be built with the correct javac once, and then simply
downloaded.

Anytime any of the custom classes change the appropriate jar version
must be bumped and the new jar along with its SHA1 should be uploaded to
s3.