Pull stopwords from package example
As upstream Solr is updated the maintainers may change the contents of
the example stopword files for the various languages supported. Rather
than tracking these in Yokozuna's source tree it makes more sense to
copy them from the Solr package. This way they are kept up to date in
sync with Solr.