Pass JVM opts to java call
A regression was made in commit
1b29749495de8e9dafbcac5b8d0d62972ac52875. The config name for the JVM
options was renamed but the code responsible for starting the JVM is
still looking for the old configuration key. Options specified in the
config are lost.

This commit patches the code to pull the correct config name and thus
appropriately pass the options to the JVM.