Remove implicit bucket association
The implicit bucket association code should have been removed in
commit e2cbdc969df8257c8ace9ce0614f60a48d27aedb. In that commit it was
decided that forming an implicit association based on the `search`
property is a bad idea. Changes were made to `yz_kv:get_index`, the
function that determines which index a bucket is associated with, but
not `yz_index:associated_buckets` which determines which buckets are
associated with an index. This commit finishes the job; making these
two related functions consistent with each other.